// src/common/enum/interest-bucket.enum.ts
export enum InterestBucket {
  CREATIVE = 'CREATIVE', // Art, Music, Design, Content
  LOGICAL = 'LOGICAL', // Maths, Science, Computers
  SOCIAL = 'SOCIAL', // Helping, Teaching, Empathy
  PRACTICAL = 'PRACTICAL', // Sports, Hands-on, Building
  LEADERSHIP = 'LEADERSHIP', // Leading, Organizing, Speaking
}